- Handle detection concurrently instead of serially.
- Introduce an intermediate channel (detectableChunksChan) and detection workers.

Some rough benchmarks running it against a handful of repos. Notice the speedup is dramatic for repos that have a lot of chunks that require detection. (ie pass the keyword and regex matches)

## trufflehog
| Status | Chunks | Bytes    | Time    |
|--------|--------|----------|---------|
| Old    | 20604  | 10716577 | 41.202s |
| New    | 20604  | 10716577 | 15.323s |
## thog
| Status | Chunks  | Bytes     | Time     |
|--------|---------|-----------|----------|
| Old    | 143843  | 237138121 | 2:15.20m |
| New    | 143843  | 237138121 | 34.784s  |
## vscode
| Status | Chunks   | Bytes     | Time     |
|--------|----------|-----------|----------|
| Old    | 1002624  | 400679800 | 2:23.19m |
| New    | 1002624  | 400679800 | 2:11.92m |
## golang
| Status | Chunks  | Bytes     | Time     |
|--------|---------|-----------|----------|
| Old    | 660366  | 298301554 | 1:12.96m |
| New    | 660366  | 298301554 | 1:13.95m |
## gitlab
| Status | Chunks   | Bytes      | Time      |
|--------|----------|------------|-----------|
| Old    | 3964241  | 1206921559 | 54:24.73m |
| New    | 3964245  | 1206924001 | 10:43.01m |

High level architecture:
![concurrent-verification drawio (2)](https://github.com/trufflesecurity/trufflehog/assets/21311841/887b565f-dc6e-4b9d-a2c0-043a2f8f9e1f)


@zricethezav Given the use of channels i'm not entirely sure if i got the `dedupeAndSend` functionality correctly implemented. Since we can no longer collect all the result chunks across all the detectors for a single chunk, I had to use a LRU cache to try and get something similar. Could you take a look and see if I totally botched it, Thanks.