@rosecodym 2023-07-31
Does mean that all of our verifiers are going to be firing (effectively) in parallel? Are we worried about the effect that might have on rate limits?