@ahrav 2023-07-31
> Does mean that all of our verifiers are going to be firing (effectively) in parallel? Are we worried about the effect that might have on rate limits?

That is a valid concern. We've been running with this config setup in enterprise for about ~5 months without any rate-limits yet. I was going to add another configuration knob to control the number of detector workers in the even we need to reduce the concurrency.