@zricethezav 2023-07-31
> @zricethezav Given the use of channels i'm not entirely sure if i got the dedupeAndSend functionality correctly implemented. Since we can no longer collect all the result chunks across all the detectors for a single chunk, I had to use a LRU cache to try and get something similar. Could you take a look and see if I totally botched it, Thanks.

It looks correct to me! nice job. One thing to note is that there is still an issue with deduping `base64` and `plain`. See https://github.com/trufflesecurity/trufflehog/issues/1465. This happens when the _secret_ gets encoded in which case the logic we have for checking deduping doesn't work... I've toyed around with some ideas like checking if `SourceMetadata` is the same and levenshtein distance is below a certain threshold.. not totally sold on that idea though.

Anyways! This PR looks good to me. Glad you could figure out how to implement this (my previous attempts failed)